Optimize second exercise
This is mainly a memory optimization. For large n's, the algorithm can
now switch to a segmented version that uses sqrt(n) memory. Also the
speed has improved and now all the primes below 1 billion can be
computed in less than 8 seconds. Note that output on small number less
than 3 is wrong.